+++ exported_sketches/encoder_router/encoder_router.ino
@@ -1,4 +1,4 @@
-// MOZZIFLOW v110.9 BALANCED CORE REFINED SKETCH
+// MOZZIFLOW v111.0 BALANCED CORE REFINED SKETCH
 #include <Mozzi.h>
 #include <Oscil.h>
 #include <tables/sin2048_int8.h>
@@ -38,19 +38,19 @@ AudioOutput updateAudio() {
         }
         arduinoencoder_enc1_last = arduinoencoder_enc1_cur;
     // Control logic moved to audio loop for node cnt1
-    if((int)node_enc1_up>0 && !mozzicounter_cnt1_ul){ mozzicounter_cnt1_c++; mozzicounter_cnt1_ul=1; } else if((int)node_enc1_up==0){ mozzicounter_cnt1_ul=0; }
-        if((int)node_enc1_down>0 && !mozzicounter_cnt1_dl){ mozzicounter_cnt1_c--; mozzicounter_cnt1_dl=1; } else if((int)node_enc1_down==0){ mozzicounter_cnt1_dl=0; }
-        if(mozzicounter_cnt1_c > (int)1) mozzicounter_cnt1_c = 0; if(mozzicounter_cnt1_c < 0) mozzicounter_cnt1_c = (int)1;
+    if((long)node_enc1_up>0 && !mozzicounter_cnt1_ul){ mozzicounter_cnt1_c++; mozzicounter_cnt1_ul=1; } else if((long)node_enc1_up==0){ mozzicounter_cnt1_ul=0; }
+        if((long)node_enc1_down>0 && !mozzicounter_cnt1_dl){ mozzicounter_cnt1_c--; mozzicounter_cnt1_dl=1; } else if((long)node_enc1_down==0){ mozzicounter_cnt1_dl=0; }
+        if(mozzicounter_cnt1_c > (long)1) mozzicounter_cnt1_c = 0; if(mozzicounter_cnt1_c < 0) mozzicounter_cnt1_c = (long)1;
         node_cnt1_out = mozzicounter_cnt1_c;
     node_osc1_out = oscil_osc1.next();
     // Control logic moved to audio loop for node osc1
-    oscil_osc1.setFreq((float)440);
+    oscil_osc1.setFreq((float)(long)440);
     // Control logic moved to audio loop for node rout1
-    node_rout1_out0 = ( (int)node_cnt1_out == 0 ) ? (int)node_osc1_out : 0;
-        node_rout1_out1 = ( (int)node_cnt1_out == 1 ) ? (int)node_osc1_out : 0;
+    node_rout1_out0 = ( (long)node_cnt1_out == 0 ) ? (int)(long)node_osc1_out : 0;
+        node_rout1_out1 = ( (long)node_cnt1_out == 1 ) ? (int)(long)node_osc1_out : 0;
     node_osc2_out = oscil_osc2.next();
     // Control logic moved to audio loop for node osc2
-    oscil_osc2.setFreq((float)220);
+    oscil_osc2.setFreq((float)(long)220);
     return MonoOutput::from8Bit((int)node_rout1_out0);
 }
